fix(wal): 🕐 extend WAL timestamp to uint64_t to prevent year-2038 overflow
The csilk_wf_wal_header_t.timestamp field used uint32_t which would overflow in 2038. Changed to uint64_t for future-proofing.

Ultraworked with Sisyphus (https://github.com/code-yeongyu/oh-my-openagent)

+++ include/csilk/app/workflow_wal.h
@@ -54,7 +54,7 @@ typedef enum {
 typedef struct __attribute__((packed)) {
 	uint32_t magic;	      /**< Must equal CSILK_WF_MAGIC for file validation. */
 	uint8_t type;	      /**< csilk_wf_event_type_t value. */
-	uint32_t timestamp;   /**< Unix timestamp (seconds since epoch) of the event. */
+	uint64_t timestamp;   /**< Unix timestamp (seconds since epoch) of the event. */
 	uint32_t payload_len; /**< Byte length of the payload that follows this header. */
 } csilk_wf_wal_header_t;
